WIP on user groups feature - Added rename group functionality - Added translations for all labels on FE - Moved the lang settings on FE to a separate service to make sure the lang settings is available from declension pipe - Added declension pipe on FE to facilitate the word casing support depending on the numeral used - Moved the CallerId field population in IAuthenticatedRequest instances from the controllers to MediatR middleware - Still need to work with contest applications

+++ contester/Features/UserGroups/Commands/DeleteUserGroupCommand.cs
@@ -1,13 +1,18 @@
+using System.Text.Json.Serialization;
+using contester.Common.MediatR;
 using contester.Features.Common.Exceptions;
+using contester.Features.UserGroups.Exceptions;
 using contester.Infrastructure.Persistence;
 using MediatR;
 using Microsoft.EntityFrameworkCore;
 
 namespace contester.Features.UserGroups.Commands;
 
-public class DeleteUserGroupCommand : IRequest<Unit>
+public class DeleteUserGroupCommand : IRequest<Unit>, IAuthorizedRequest
 {
     public Guid GroupId { get; set; }
+    [JsonIgnore] public Guid CallerId { get; set; }
+    [JsonIgnore] public Constants.Permission RequiredPermission => Constants.Permission.ManageContestParticipants;
 }
 
 public class DeleteUserGroupCommandHandler(
@@ -17,7 +22,10 @@ public class DeleteUserGroupCommandHandler(
     {
         var group = await context.UserGroups.FirstOrDefaultAsync(ug => ug.Id == request.GroupId, ct);
         if (group == null)
-            throw new NotifyUserException("Group not found");
+            throw new EntityNotFoundException(typeof(UserGroup), request.GroupId);
+
+        if (group.IsSystemGroup)
+            throw new CannotDropSystemGroupException();
         
         context.UserGroups.Remove(group);
